requirements.txt
Django==1.11.11
Pillow==4.3.0                   # PIL fork (Python Imaging Library)
django-tastypie==0.14.0         # API v0
djangorestframework==3.7.7      # API v1
django-filter==1.1.0            # Filtering for DRF
# python-memcached==1.57          # Memcache. Used by Mailinglists to fetch from Sympa.
# Upstream is missing Python 3 patches
git+https://github.com/JelteF/python-memcached@patch-1#egg=python-memcached==1.58
markdown2==2.3.5                # textarea text formatting
pytz                            # Timezone lib
stripe==1.79.1                  # Stripe payment
icalendar==4.0.1                # iCalendar generation
google-api-python-client==1.6.5

# third party deps
django-filebrowser==3.9.1       # File uploading
git+https://github.com/dotKom/django-chunks.git@e6bf109d4fd286964024c140ac1b8f3d86713540#egg=django-chunks==0.4
django-crispy-forms==1.7.2      # nice forms
django-extensions==1.9.7        # nice extra commands for debugging, etc
django-dynamic-fixture==1.9.5   # Dynamic fixtures for models
django-recaptcha==1.4.0         # Google ReCaptcha
django-oauth-toolkit==1.0.0    # OAuth2 authentication support
django-watson==1.5.2            # Indexed model search lib
django-reversion==2.0.13        # Model version history with middleware hooks to all post_save signals
git+https://github.com/dotkom/django-guardian.git@1b184d5377200ab4ad59b343ae8af9a863124510#egg=django-guardian==1.4.9.1          # Per Object permissions framework
django-taggit==0.22.2           # Generic multi-model tagging library
django-taggit-serializer==0.1.5 # REST Framework serializers for Django-taggit
APScheduler==3.5.1              # Scheduler
feedme==1.9.5
git+https://github.com/dotKom/redwine.git@0690807537fcbeeacde9c40b0e45549e0d216a96#egg=redwine==1.2.4.1
reportlab==3.4.0
pdfdocument==3.2
Unidecode==0.04.21               # Translates every unicode symbol to the closest ascii. online_mail_usernames
django-markdown-deux==1.0.5
django-formset-js==0.5.0        # used to dynamically create new formsets
django-cors-headers==2.2.0      # Enable CORS for Nibble
django-datetime-widget==0.9.3   # Datetime widget for forms
django-webpack-loader==0.6.0    # Integration with webpack
dj-database-url==0.5.0          # Allows to configure databases using URLs
python-decouple==3.1            # Configuration and settings management
django-oidc-provider==0.5.3     # OpenID Connect Provider
docutils==0.14                  # Allows django-admindocs to generate documentation
graphene-django==2.0

#test tools
django-nose>=1.4,<1.5           # We use this test runner locally
nose==1.3.7                     # We use this test runner locally

# Frigg
requests[security]==2.18.4

# Wiki
wiki==0.3.1

# Django 1.9 support
sorl-thumbnail>12.2
django-appconf>=1.0

# non-pip
# imagemagick                   # From package repository. Tested with version == 8:6.7.7.10.
# libjpeg-dev					# From package repository.
# libpq-dev						# From package repository.

# Monitoring
raven==6.6.0

django-js-reverse==0.8.1
